back-end/requirements.txt
flask
flask-pymongo
flask-cors
python-dotenv
pyjwt
werkzeug
razorpay